+++ playground/src/router.ts
@@ -11,6 +11,16 @@ const routes = [
         name: 'Multiple',
         component: () => import('./pages/multiple.vue'),
     },
+    {
+        path: '/multiple-cameras',
+        name: 'Multiple Cameras',
+        component: () => import('./pages/multiple-cameras.vue'),
+    },
+    {
+        path: '/no-camera',
+        name: 'No Camera',
+        component: () => import('./pages/no-camera.vue'),
+    },
 ]
 export const router = createRouter({
     history: createWebHistory(),