+++ src/composables/usePointerEventHandler/index.ts
@@ -1,19 +1,28 @@
 import { uniqueBy } from '../../utils'
 import { useRaycaster } from '../useRaycaster'
 import { computed, reactive } from 'vue'
-import type { Intersection, Event, Object3D } from 'three'
 
-type CallbackFn = (intersection: Intersection<Object3D<Event>>, event: PointerEvent) => void //TODO document
+import type { TresContext } from '../useTresContextProvider'
+import type { Intersection, Event, Object3D } from 'three'
+import { TresScene } from 'src/types'
+
+type CallbackFn = (intersection: Intersection<Object3D<Event>>, event: PointerEvent) => void
 type CallbackFnPointerLeave = (object: Object3D<Event>, event: PointerEvent) => void
 
-type EventProps = {
+export type EventProps = {
   onClick?: CallbackFn
   onPointerEnter?: CallbackFn
   onPointerMove?: CallbackFn
   onPointerLeave?: CallbackFnPointerLeave
 }
 
-export const usePointerEventHandler = () => {
+export const usePointerEventHandler = (
+  { scene, contextParts }:
+    {
+      scene: TresScene,
+      contextParts: Pick<TresContext, 'renderer' | 'camera'>
+    }
+) => {
   const objectsWithEventListeners = reactive({
     click: new Map<Object3D, CallbackFn>(),
     pointerMove: new Map<Object3D, CallbackFn>(),
@@ -32,15 +41,12 @@ export const usePointerEventHandler = () => {
     if (onPointerMove) objectsWithEventListeners.pointerMove.set(object, onPointerMove)
     if (onPointerEnter) objectsWithEventListeners.pointerEnter.set(object, onPointerEnter)
     if (onPointerLeave) objectsWithEventListeners.pointerLeave.set(object, onPointerLeave)
-
-    object.addEventListener('removed', () => {
-      object.traverse((child: Object3D) => {
-        deregisterObject(child)
-      })
-
-      deregisterObject(object)
-    })
   }
+
+  // to make the registerObject available in the custom renderer (nodeOps), it is attached to the scene
+  scene.userData.tres__registerAtPointerEventHandler = registerObject
+  scene.userData.tres__deregisterAtPointerEventHandler = deregisterObject
+
 
   const objectsToWatch = computed(() =>
     uniqueBy(
@@ -51,7 +57,7 @@ export const usePointerEventHandler = () => {
     ),
   )
 
-  const { onClick, onPointerMove } = useRaycaster(objectsToWatch)
+  const { onClick, onPointerMove } = useRaycaster(objectsToWatch, contextParts)
 
   onClick(({ intersects, event }) => {
     if (intersects.length) objectsWithEventListeners.click.get(intersects[0].object)?.(intersects[0], event)