+++ src/keys.ts
@@ -1,7 +1,0 @@
-export const UseTresStateSymbol = Symbol('UseTresState')
-
-export const OBJECT_3D_USER_DATA_KEYS = {
-  GEOMETRY_VIA_PROP: 'tres__geometryViaProp',
-  MATERIAL_VIA_PROP: 'tres__materialViaProp',
-  REGISTER_AT_POINTER_EVENT_HANDLER: 'tres__registerAtPointerEventHandler',
-}